refactor(charts): relocate shared dataset types to a neutral module
Move IDatasetServiceDatapoint, TimeScaleFormat, IDatasetServiceDatasetConfig,
and IDatasetServiceDataSourceInfo out of dataset-stream.service.ts into
core/interfaces/dataset.interfaces.ts, and re-point every importer. These
contracts are shared by the History engine, config, and persistence layers,
so they must outlive the recorder; relocating them lets a later PR delete
DatasetStreamService without a compile cascade. Pure move, no behavior change.

Part of #64.
